refactor: 抽离通用无损JSON解析工具并整理代码
1.  提取重复的lossless-json解析逻辑到json-utils.ts
2.  修复tooltip打开条件判断逻辑和注释
3.  替换所有文件内联的JSON解析代码
4.  更新脚本注释说明和文件结构
5.  统一替换双引号为单引号并删除旧的统计翻译文件
