+++ scripts/lib/convert-icons.ts
@@ -4,7 +4,7 @@
 // ================================================================================
 // Parameters: quality 50, chroma 4:2:0, effort 4
 
-import { existsSync, readdirSync, mkdirSync, readFileSync } from 'node:fs'
+import { existsSync, readdirSync, statSync, mkdirSync, writeFileSync, readFileSync } from 'node:fs'
 import { join } from 'node:path'
 import sharp from 'sharp'
 
@@ -16,9 +16,12 @@ export interface IconConversionResult {
   missingSource: string[]
 }
 
+const CATEGORY_PREFIX: Record<string, string> = { weapon: 'wpn_', equip: 'item_equip_' }
+
 /**
  * Convert only the specified PNG icons from AKEDatabase to AVIF.
  * `targetIds` is a list of icon filenames WITHOUT extension (e.g. "wpn_sword_0017").
+ * Skips conversion when AVIF already exists and is not older than the source PNG.
  * Both weapon and equip icons are in iconbig/ under their respective category dirs.
  */
 export async function convertIcons(
@@ -31,16 +34,15 @@ export async function convertIcons(
   const targetSet = new Set(targetIds)
 
   for (const category of categories) {
-    // Both weapon and equip large icons are in iconbig/
     const srcDir = join(akedatabasePath, 'public', 'images', category, 'iconbig')
     const outDir = join(projectPublicDir, 'images', category)
     mkdirSync(outDir, { recursive: true })
 
     if (!existsSync(srcDir)) {
       console.warn(`  [icons] source dir not found: ${srcDir}`)
-      // Mark all targets for this category as missing source
+      const prefix = CATEGORY_PREFIX[category]
       for (const id of targetIds) {
-        if (id.startsWith(category === 'weapon' ? 'wpn_' : 'item_equip_')) {
+        if (id.startsWith(prefix)) {
           result.missingSource.push(`${category}/${id}.png`)
         }
       }
@@ -55,9 +57,14 @@ export async function convertIcons(
       const srcPath = join(srcDir, file)
       const outPath = join(outDir, file.replace('.png', '.avif'))
 
+      // Skip if AVIF already exists and is not older than source PNG
       if (existsSync(outPath)) {
-        result.skipped.push(file)
-        continue
+        const srcTime = statSync(srcPath).mtimeMs
+        const outTime = statSync(outPath).mtimeMs
+        if (outTime >= srcTime) {
+          result.skipped.push(file)
+          continue
+        }
       }
 
       try {
@@ -65,7 +72,6 @@ export async function convertIcons(
         const avifBuffer = await sharp(pngBuffer)
           .avif(AVIF_OPTIONS)
           .toBuffer()
-        const { writeFileSync } = await import('node:fs')
         writeFileSync(outPath, avifBuffer)
         result.converted.push(file)
       } catch (err) {
